Refactor: Move utility scripts to 'tools' directory
Reorganized utility scripts from 'utils' to 'tools' for improved clarity and structure. Updated imports, file references, and relevant paths across the codebase to reflect this change. Adjusted scripts in `package.json` and a GitHub workflow configuration accordingly.
