+++ .github/workflows/check-missing-localizations.yml
@@ -25,7 +25,7 @@ jobs:
 
       - name: Run localization checker
         id: localization_checker
-        run: node utils/find-missing-localizations.mjs
+        run: node tools/find-missing-localizations.mjs
           --output-file "./reports/localization/missing-keys-report.json"
         continue-on-error: true
