+++ gulpfile.mjs
@@ -1,7 +1,7 @@
 import gulp from "gulp";
 
-import * as css from "./utils/css.mjs";
-import * as javascript from "./utils/javascript.mjs";
+import * as css from "./tools/css.mjs";
+import * as javascript from "./tools/javascript.mjs";
 
 
 // default export - build CSS and watch for updates
