+++ tools/consolidate-lang-json.mjs
