+++ tools/format-lang-json.mjs
